# Docs/api-docs.yaml
openapi: 3.0.1
info:
  title: Event Signup API
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
  /update:
    put:
      tags:
      - form-controller
      operationId: updateForm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
  /api/v1/event/update:
    put:
      tags:
      - event-controller
      summary: Update an existing event.
      operationId: updateEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized.
        "200":
          description: Event updated successfully.
  /create:
    post:
      tags:
      - form-controller
      operationId: createForm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
  /api/v1/signup/{eventId}/add:
    post:
      tags:
      - signup-controller
      operationId: signupToAnEvent
      parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParticipantDto'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
  /api/v1/event/create:
    post:
      tags:
      - event-controller
      summary: Create a new event.
      operationId: createNewEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEventAndFormRequest'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Event creation successful.
        "401":
          description: Unauthorized.
...
  /api/v1/signup/{eventId}:
    get:
      tags:
      - signup-controller
      summary: Get an event for signup purposes.
      operationId: fetchSignupEvent
      parameters:
      - name: eventId
        in: path
        description: Event's id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "404":
          description: Event was already held.
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Signup not started/signup already ended/event full. See the
            message in response for details.
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: The event requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDto'
  /api/v1/event/users-events:
    get:
      tags:
      - event-controller
      summary: Get logged in user's events and their forms.
      operationId: fetchUsersEvents
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Requested data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersEvents'
        "401":
          description: Unauthorized.
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UsersEvents'
  /api/v1/event/events:
    get:
      tags:
      - event-controller
      summary: "Get latest, upcoming and past events."
      operationId: fetchEvents
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "406":
          description: Not Acceptable
          content:
            '*/*':
              schema:
                type: string
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized.
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventsDto'
        "200":
          description: Requested data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsDto'
...
components:
  schemas:
    FormDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        userEmail:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FormField'
    FormField:
      type: object
      properties:
        order:
          type: integer
          format: int32
        label:
          type: string
        type:
          type: string
        value:
          type: array
          items:
            type: string
    Answer:
      type: object
      properties:
        order:
          type: integer
          format: int32
        label:
          type: string
        value:
          type: string
    EventDto:
      required:
      - description
      - formId
      - name
      - place
      - signupStarts
      - startDate
      type: object
      properties:
        id:
          type: string
          format: uuid
        formId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        place:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        minParticipants:
          type: integer
          format: int32
        maxParticipants:
          type: integer
          format: int32
        signupStarts:
          type: string
          format: date-time
        signupEnds:
          type: string
          format: date-time
        price:
          type: number
          format: double
        bannerImg:
          type: string
        metaData:
          type: object
          additionalProperties:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantDto'
    ParticipantDto:
      required:
      - eventId
      - userEmail
      type: object
      properties:
        id:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        userEmail:
          type: string
        answers:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Answer'
    NewEventAndFormRequest:
      type: object
      properties:
        event:
          $ref: '#/components/schemas/EventDto'
        form:
          $ref: '#/components/schemas/FormDto'
...
    UsersEvents:
      type: object
      properties:
        myEvents:
          type: array
          items:
            $ref: '#/components/schemas/EventDto'
        myForms:
          type: array
          items:
            $ref: '#/components/schemas/FormDto'
    EventsDto:
      type: object
      properties:
        myEvents:
          $ref: '#/components/schemas/MyEvents'
        myForms:
          $ref: '#/components/schemas/MyForms'
    MyEvents:
      type: object
      properties:
        newEvents:
          type: array
          items:
            $ref: '#/components/schemas/EventDto'
        upcomingEvents:
          type: array
          items:
            $ref: '#/components/schemas/EventDto'
        pastEvents:
          type: array
          items:
            $ref: '#/components/schemas/EventDto'
    MyForms:
      type: object
      properties:
        newEventsForms:
          type: array
          items:
            $ref: '#/components/schemas/FormDto'
        upcomingEventsForms:
          type: array
          items:
            $ref: '#/components/schemas/FormDto'
        pastEventsForms:
          type: array
          items:
            $ref: '#/components/schemas/FormDto'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT